requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.0.350
langchain-google-genai==0.0.5
google-genai==0.1.0
pgvector==0.2.4
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
uuid==1.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 